Move table properties ui utility methods to utils/ui/table-properties.js.
