+++ packages/ckeditor5-table/src/tablecellproperties/tablecellpropertiesui.js
@@ -20,7 +20,7 @@ import {
 	defaultColors,
 	lengthFieldValidator,
 	lineWidthFieldValidator
-} from '../ui/utils';
+} from '../utils/ui/table-properties';
 import {
 	getLocalizedColorOptions,
 	normalizeColorOptions
